+++ dbvisitor-core/src/main/java/net/hasor/dbvisitor/dialect/provider/Oracle12cDialect.java
@@ -34,7 +34,7 @@ public class Oracle12cDialect extends OracleDialect {
     }
 
     @Override
-    public BoundSql pageSql(BoundSql boundSql, int start, int limit) {
+    public BoundSql pageSql(BoundSql boundSql, long start, long limit) {
         StringBuilder sqlBuilder = new StringBuilder(boundSql.getSqlString());
         List<Object> paramArrays = new ArrayList<>(Arrays.asList(boundSql.getArgs()));
